% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_copy}
\alias{sp_copy}
\title{Copy a matrix using armadillo's iterators}
\usage{
sp_copy(mat)
}
\arguments{
\item{mat}{dgCMatrix. Matrix to copy.}
}
\description{
Copy a matrix using armadillo's iterators
}

