% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_runif}
\alias{sp_runif}
\title{Generate random uniform matrix using armadillo}
\usage{
sp_runif(nrows, ncols, d)
}
\arguments{
\item{nrows}{Integer scalar. Number of rows.}

\item{ncols}{Integer scalar. Number of columns.}

\item{d}{Numeric scalar. Density.}
}
\description{
Generate random uniform matrix using armadillo
}

